Make ChunkLoadPadding a float like ChunkLoadRadius

+++ engine.go
@@ -4,7 +4,7 @@ var (
 	ChunksDir              string
 	ChunkSize              int
 	ChunkLoadRadius        float64
-	ChunkLoadPadding       int
+	ChunkLoadPadding       float64
 	WorldMaxAltitude       int
 	GWorld                 World
 	TileSize               int
